Оформление классов, реализован функционал получения имени трекера из текст филд, наведен порядок в протоколах
