+++ Tracker/Helpers/UITextField+Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  UITextField+Extensions.swift
+//  Tracker
+//
+//  Created by Anastasia Gorbunova on 10.03.2024.
+//
+
+import Foundation
+import UIKit
+
+extension UITextField {
+    func setLeftPaddingPoints(_ amount:CGFloat){
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+    func setRightPaddingPoints(_ amount:CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
+        self.rightView = paddingView
+        self.rightViewMode = .always
+    }
+}
