+++ Tracker/Schedule/ScheduleProtocol.swift
@@ -9,5 +9,4 @@ import Foundation
 
 protocol ScheduleProtocol: AnyObject {
     func saveSelectedDays(selectedDays: Set<WeekDays>)
-    //func showScheduleViewController(viewController: ScheduleViewController)
 }
