+++ Tracker/Schedule/ScheduleTableCell.swift
@@ -15,19 +15,16 @@ final class ScheduleTableCell: UITableViewCell {
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-        backgroundColor = UIColor(named: "YP Gray")?.withAlphaComponent(0.3)
+        backgroundColor = .ypLightGray.withAlphaComponent(0.3)
         setUpSwitch()
-        //layer.masksToBounds = true
-        
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
     private func setUpSwitch() {
-        switchButton.setOn(false, animated: true)
-        
+        switchButton.setOn(false, animated: true)      
         self.accessoryView = switchButton
-        
     }
 }
