+++ Tracker/Schedule/ScheduleViewController.swift
@@ -11,9 +11,10 @@ import UIKit
 final class ScheduleViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
     
     weak var sheduleDelegate: ScheduleProtocol?
+    var selectedDays: Set<WeekDays> = []
+    
     private var tableView = UITableView()
     private let saveButton = UIButton()
-    var selectedDays: Set<WeekDays> = []
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -25,46 +26,30 @@ final class ScheduleViewController: UIViewController, UITableViewDataSource, UIT
         initTableView()
     }
     
+    //MARK: - Data Source
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        7
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: ScheduleTableCell.identifier, for: indexPath) as? ScheduleTableCell else  {
+            return UITableViewCell()
+        }
+        configureCell(cell: cell, cellForRowAt: indexPath)
+        return cell
+    }
+    
+    //MARK: - Delegate
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 75
+    }
+    
+    //MARK: - Actions
     @objc
     private func saveButtonPressed() {
         navigationController?.popViewController(animated: true)
         sheduleDelegate?.saveSelectedDays(selectedDays: selectedDays)
         
-    }
-    
-    private func setUpSaveButton() {
-        saveButton.setTitle("Готово", for: .normal)
-        saveButton.backgroundColor = UIColor(named: "YP Black") ?? .black
-        saveButton.layer.cornerRadius = 16
-        saveButton.addTarget(self, action: #selector(saveButtonPressed), for: .touchUpInside)
-        
-        view.addSubview(saveButton)
-        saveButton.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            saveButton.heightAnchor.constraint(equalToConstant: 60),
-            saveButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            saveButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            saveButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16)
-        ])
-    }
-    
-    private func initTableView() {
-        tableView.register(ScheduleTableCell.self, forCellReuseIdentifier: ScheduleTableCell.identifier)
-        
-        tableView.dataSource = self
-        tableView.delegate = self
-        
-        tableView.separatorInset = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
-        tableView.tableHeaderView = UIView()
-        
-        view.addSubview(tableView)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 16),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            tableView.heightAnchor.constraint(equalToConstant: 525),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16)
-        ])
     }
     
     @objc
@@ -80,6 +65,25 @@ final class ScheduleViewController: UIViewController, UITableViewDataSource, UIT
             }
         }
     }
+    
+    //MARK: - Private Methods
+    
+    private func setUpSaveButton() {
+        saveButton.setTitle("Готово", for: .normal)
+        saveButton.backgroundColor = .ypBlack
+        saveButton.layer.cornerRadius = 16
+        saveButton.addTarget(self, action: #selector(saveButtonPressed), for: .touchUpInside)
+        
+        view.addSubview(saveButton)
+        saveButton.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            saveButton.heightAnchor.constraint(equalToConstant: 60),
+            saveButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
+            saveButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
+            saveButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16)
+        ])
+    }
+    
     private func configureCell(cell: ScheduleTableCell, cellForRowAt indexPath: IndexPath) {
         guard let weekday = WeekDays(rawValue: indexPath.row + 1) else {
             return
@@ -102,21 +106,25 @@ final class ScheduleViewController: UIViewController, UITableViewDataSource, UIT
             cell.switchButton.setOn(true, animated: true)
         }
     }
-    //MARK: - Data Source
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        7
-    }
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: ScheduleTableCell.identifier, for: indexPath) as? ScheduleTableCell else  {
-            return UITableViewCell()
-        }
-        configureCell(cell: cell, cellForRowAt: indexPath)
-        return cell
-    }
+    //MARK: - Table Initialization
     
-    //MARK: - Delegate
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 75
+    private func initTableView() {
+        tableView.register(ScheduleTableCell.self, forCellReuseIdentifier: ScheduleTableCell.identifier)
+        
+        tableView.dataSource = self
+        tableView.delegate = self
+        
+        tableView.separatorInset = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
+        tableView.tableHeaderView = UIView()
+        
+        view.addSubview(tableView)
+        tableView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 16),
+            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
+            tableView.heightAnchor.constraint(equalToConstant: 525),
+            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16)
+        ])
     }
 }
