+++ Tracker/TrackerCreation/Cells/ColorCell.swift
@@ -18,6 +18,7 @@ final class ColorCell: UICollectionViewCell {
     
     let colorView = UIView()
     static let identifier = "ColorCell"
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
         self.layer.cornerRadius = 8
@@ -32,11 +33,9 @@ final class ColorCell: UICollectionViewCell {
             colorView.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
             colorView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
         ])
-        
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
 }
