+++ Tracker/TrackerCreation/Cells/EmojiCell.swift
@@ -11,8 +11,7 @@ import UIKit
 final class EmojiCell: UICollectionViewCell {
     override var isSelected: Bool {
         didSet {
-            self.backgroundColor = self.isSelected ? (UIColor(named: "YP Gray") ?? .gray) : .clear
-            
+            self.backgroundColor = self.isSelected ? UIColor.ypLightGray : .clear
         }
     }
     
@@ -21,12 +20,10 @@ final class EmojiCell: UICollectionViewCell {
     override init(frame: CGRect) {
         super.init(frame: frame)
         
-        //contentView.backgroundColor = .gray
         label.font = UIFont.systemFont(ofSize: 32, weight: .bold)
         contentView.addSubview(label)
         label.translatesAutoresizingMaskIntoConstraints = false
         self.layer.cornerRadius = 16
-        
         
         NSLayoutConstraint.activate([                                    // 5
             label.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
@@ -37,10 +34,5 @@ final class EmojiCell: UICollectionViewCell {
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setUpBackground() {
-        self.layer.cornerRadius = 16
-        //self.backgroundColor = .green
     }
 }
