+++ Tracker/TrackerCreation/Cells/NameTrackerCell.swift
@@ -8,57 +8,47 @@
 import Foundation
 import UIKit
 
-extension UITextField {
-    func setLeftPaddingPoints(_ amount:CGFloat){
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.leftView = paddingView
-        self.leftViewMode = .always
-    }
-    func setRightPaddingPoints(_ amount:CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.rightView = paddingView
-        self.rightViewMode = .always
-    }
+protocol SaveNameTrackerDelegate: AnyObject {
+    func textFieldWasChanged(text: String)
 }
 
 final class NameTrackerCell: UICollectionViewCell {
     
     static let identifier = "TrackerNameTextFieldCell"
+    weak var delegate: SaveNameTrackerDelegate?
     
-    var trackerNameTextField = UITextField()
+    let trackerNameTextField = UITextField()
     let xButton = UIButton(type: .custom)
     
     override init(frame: CGRect) {
         super.init(frame: frame)
         
-        
         setUpNameTextField()
-        //setUpButton()
-        
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-//    private func setUpButton() {
-//        let image = UIImage(systemName: "multiply.circle.fill")
-//        xButton.setImage(image, for: .normal)
-//
-//        trackerNameTextField.rightView = xButton
-//        trackerNameTextField.rightViewMode = .always
-//        trackerNameTextField.setRightPaddingPoints(16)
-//    }
+    @objc
+    func textFieldEditingChanged(_ textField: UITextField) {
+        guard let text = trackerNameTextField.text else {
+            return
+        }
+        delegate?.textFieldWasChanged(text: text)
+    }
     
     private func setUpNameTextField() {
         contentView.addSubview(trackerNameTextField)
         trackerNameTextField.layer.cornerRadius = 16
-        trackerNameTextField.backgroundColor = UIColor(named: "YP Gray")?.withAlphaComponent(0.3) ?? .gray.withAlphaComponent(0.3)
+        trackerNameTextField.backgroundColor = UIColor.ypLightGray.withAlphaComponent(0.3)
         trackerNameTextField.placeholder = "Введите название трекера"
         trackerNameTextField.font = UIFont.systemFont(ofSize: 17, weight: .regular)
         trackerNameTextField.setLeftPaddingPoints(12)
         
-        //trackerNameTextField.becomeFirstResponder()
+        trackerNameTextField.clearButtonMode = .whileEditing
+        
+        trackerNameTextField.addTarget(self, action: #selector(textFieldEditingChanged(_:)), for: .editingDidEnd)
         
         trackerNameTextField.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
